requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SpeechRecognition==3.10.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydub==0.25.1
